move rule to core
